punt conversion logic outside of the database
