+++ Content.IntegrationTests/Tests/Preferences/ServerDbSqliteTests.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using Content.Server.Database;
+using Content.Server.Preferences.Managers;
 using Content.Shared.GameTicking;
 using Content.Shared.Humanoid;
 using Content.Shared.Humanoid.Prototypes;
@@ -61,13 +62,12 @@ namespace Content.IntegrationTests.Tests.Preferences
         {
             var cfg = server.ResolveDependency<IConfigurationManager>();
             var serialization = server.ResolveDependency<ISerializationManager>();
-            var task = server.ResolveDependency<ITaskManager>();
             var opsLog = server.ResolveDependency<ILogManager>().GetSawmill("db.ops");
             var builder = new DbContextOptionsBuilder<SqliteServerDbContext>();
             var conn = new SqliteConnection("Data Source=:memory:");
             conn.Open();
             builder.UseSqlite(conn);
-            return new ServerDbSqlite(() => builder.Options, true, cfg, true, opsLog, task, serialization);
+            return new ServerDbSqlite(() => builder.Options, true, cfg, true, opsLog, serialization);
         }
 
         [Test]
@@ -86,12 +86,14 @@ namespace Content.IntegrationTests.Tests.Preferences
         {
             var pair = await PoolManager.GetServerClient();
             var db = GetDb(pair.Server);
+            var preferences = (ServerPreferencesManager)pair.Server.ResolveDependency<IServerPreferencesManager>();
             var username = new NetUserId(new Guid("640bd619-fc8d-4fe2-bf3c-4a5fb17d6ddd"));
             const int slot = 0;
             var originalProfile = CharlieCharlieson();
             await db.InitPrefsAsync(username, originalProfile);
             var prefs = await db.GetPlayerPreferencesAsync(username);
-            Assert.That(prefs.Characters.Single(p => p.Key == slot).Value.MemberwiseEquals(originalProfile));
+            var profile = preferences.ConvertProfiles(prefs!.Profiles.Find(p => p.Slot == slot));
+            Assert.That(profile.MemberwiseEquals(originalProfile));
             await pair.CleanReturnAsync();
         }
 
@@ -107,7 +109,7 @@ namespace Content.IntegrationTests.Tests.Preferences
             await db.SaveSelectedCharacterIndexAsync(username, 1);
             await db.SaveCharacterSlotAsync(username, null, 1);
             var prefs = await db.GetPlayerPreferencesAsync(username);
-            Assert.That(!prefs.Characters.Any(p => p.Key != 0));
+            Assert.That(prefs!.Profiles, Has.Count.EqualTo(1));
             await pair.CleanReturnAsync();
         }
 
@@ -138,6 +140,7 @@ namespace Content.IntegrationTests.Tests.Preferences
             var pair = await PoolManager.GetServerClient();
             var server = pair.Server;
             var db = GetDb(pair.Server);
+            var preferences = (ServerPreferencesManager)pair.Server.ResolveDependency<IServerPreferencesManager>();
 
             var proto = server.ResolveDependency<IPrototypeManager>();
             Assert.That(!proto.HasIndex<SpeciesPrototype>(InvalidSpecies), "You should not have added a species called WingusDingus, but change it in this test to something else I guess");
@@ -155,21 +158,17 @@ namespace Content.IntegrationTests.Tests.Preferences
             if (legacy)
                 await db.MakeCharacterSlotLegacyAsync(username, 0);
 
-            await server.WaitIdleAsync();
-
-            Task<PlayerPreferences> task = null!;
-            PlayerPreferences prefs = null;
-            await server.WaitPost(() =>
-            {
-                task = db.GetPlayerPreferencesAsync(username, CancellationToken.None);
-            });
-            await pair.RunTicksSync(20);
-            Assert.DoesNotThrowAsync(async () => prefs = await task);
+            var prefs = await db.GetPlayerPreferencesAsync(username, CancellationToken.None);
 
             Assert.That(prefs, Is.Not.Null);
-            Assert.That(prefs.Characters, Has.Count.EqualTo(1));
-            Assert.That(prefs.Characters[0].Species, Is.Not.EqualTo(InvalidSpecies));
-            Assert.That(prefs.Characters[0].Species, Is.EqualTo(HumanoidCharacterProfile.DefaultSpecies));
+            await server.WaitAssertion(() =>
+            {
+                var converted = preferences.ConvertPreferences(prefs);
+
+                Assert.That(converted.Characters, Has.Count.EqualTo(1));
+                Assert.That(converted.Characters[0].Species, Is.Not.EqualTo(InvalidSpecies));
+                Assert.That(converted.Characters[0].Species, Is.EqualTo(HumanoidCharacterProfile.DefaultSpecies));
+            });
 
             await pair.CleanReturnAsync();
         }
